refactor(folders, users, items, schedules): 화면별 View 패키지 파일 나누기
