+++ src/main/java/com/toit/folders/Folders.java
@@ -1,6 +1,7 @@
-package com.toit.entity;
+package com.toit.folders;
 
-import com.toit.enums.EntityStatus;
+import com.toit.common.enums.EntityStatus;
+import com.toit.user.Users;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -10,32 +11,43 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import org.springframework.data.annotation.CreatedDate;
 
+
 @Entity
-public class Schedules {
+@Table(name = "folders")
+public class Folders {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long schedulesId;
+    private Long foldersId;
 
     /**
-     * 스케줄 제목
+     * 보관함 이름
+     * 길이 제한 50
+     */
+    @Column(nullable = false, length = 50)
+    private String name;
+
+    /**
+     * 보관함 메모
+     * 있어도 되고, 없어도 됨
      * 길이 제한 255
      */
-    @Column(nullable = false)
-    private String title;
+    @Column(nullable = true)
+    private String memo;
 
     /**
-     * 스케줄 일정의 시작 날짜 및 시간
+     * 보관함 기본 보관함 여부
+     * 기본 보관함 이면 1, 기본 보관함이 아니면 0
      */
-    @Column(nullable = true)
-    private LocalDateTime startAt;
+    private Boolean isDefault;
 
     /**
-     * 스케줄 일정의 종료 날짜 및 시간
+     * 보관함 색깔 지정 -> #FFAAOO UI용 색상값
      */
-    @Column(nullable = true)
-    private LocalDateTime endAt;
+    @Column(length = 20)
+    private String color;
 
     /**
      * 보관함 서비스 상태
@@ -46,23 +58,29 @@ public class Schedules {
     private EntityStatus status;
 
     /**
-     * 스케줄 생성 후 시간 즉시 생성
+     * 보관함 생성 후 시간 즉시 생성
      */
     @CreatedDate
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
 
     /**
-     * 스케줄 서비스 상태가 DELETED 되면 업데이트
+     * 보관함 서비스 상태가 DELETED 되면 업데이트
      */
     @Column(nullable = true)
     private LocalDateTime deletedAt;
 
     /**
-     * Users와 N:1 관계 설정
+     * 보관함 즐겨찾기 기능
+     * 즐겨찾기 설정 -> 1, 즐겨찾기 비설정 -> 0
      */
+    private Boolean isFavorite;
+
+
     @ManyToOne
     @JoinColumn(name = "users_id", nullable = false)
     private Users users;
+
+
 
 }
