+++ src/main/java/com/toit/user/UsersController.java
@@ -1,0 +1,4 @@
+package com.toit.user;
+
+public class UsersController {
+}
