+++ src/main/java/com/toit/user/UsersService.java
@@ -1,0 +1,4 @@
+package com.toit.user;
+
+public class UsersService {
+}
