+++ src/main/java/com/toit/view/folders/FoldersUseCaseImpl.java
@@ -1,0 +1,4 @@
+package com.toit.view.folders;
+
+public class FoldersUseCaseImpl {
+}
