+++ src/main/java/com/toit/view/folders/FoldersViewController.java
@@ -1,0 +1,4 @@
+package com.toit.view.folders;
+
+public class FoldersViewController {
+}
