+++ src/main/java/com/toit/view/folders/dto/FoldersViewResponse.java
@@ -1,0 +1,4 @@
+package com.toit.view.folders.dto;
+
+public class FoldersViewResponse {
+}
