+++ src/main/java/com/toit/view/home/HomeUseCase.java
@@ -1,0 +1,4 @@
+package com.toit.view.home;
+
+public class HomeUseCase {
+}
