+++ src/main/java/com/toit/view/home/dto/HomeViewResponse.java
@@ -1,0 +1,4 @@
+package com.toit.view.home.dto;
+
+public class HomeViewResponse {
+}
